+++ lib/ui/health/Covid19SymptomsPanel.dart
@@ -141,7 +141,7 @@ class _Covid19SymptomsPanelState extends State<Covid19SymptomsPanel> {
   Widget _buildSymptom(HealthSymptom symptom) {
     bool _selected = _selectedSymptoms.contains(symptom.id);
     String imageName = _selected ? 'images/icon-selected-checkbox.png' : 'images/icon-deselected-checkbox.png';
-    String symptomName = _rules.localeString(symptom.name);
+    String symptomName = (_rules?.localeString(symptom?.name) ?? symptom?.name) ?? '';
     return Semantics(
       label: symptomName,
       value: (_selected?Localization().getStringEx("toggle_button.status.checked", "checked",) :
@@ -241,7 +241,8 @@ class _Covid19SymptomsPanelState extends State<Covid19SymptomsPanel> {
 
         _selectedSymptoms.add(symptom.id);
       }
-      AppSemantics.announceCheckBoxStateChange(context, _selectedSymptoms?.contains(symptom.id), _rules.localeString(symptom.name));
+      String symptomName = (_rules?.localeString(symptom?.name) ?? symptom?.name) ?? '';
+      AppSemantics.announceCheckBoxStateChange(context, _selectedSymptoms?.contains(symptom.id), symptomName);
     });
   }
 
@@ -260,7 +261,7 @@ class _Covid19SymptomsPanelState extends State<Covid19SymptomsPanel> {
       _submittingSymptoms = true;
     });
 
-    Health().processSymptoms(groups: _symptomsGroups, selected: _selectedSymptoms).then((dynamic result) {
+    Health().processSymptoms(rules: _rules, selected: _selectedSymptoms).then((dynamic result) {
       if (mounted) {
         setState(() {
           _submittingSymptoms = false;